requirements.txt
Django>=4.0
requests>=2.25.0
django-turnstile>=0.2.0